Remove unnecessary dependency from node in blockchain module
