+++ exonum/src/blockchain/service.rs
@@ -182,84 +182,6 @@ pub trait ServiceContext {
     fn add_transaction(&mut self, tx: Box<Transaction>);
 }
 
-#[doc(hidden)]
-pub struct NodeHandlerContext<'a, 'b> {
-    state: &'a mut State,
-    snapshot: &'b Snapshot,
-    txs: Vec<Box<Transaction>>,
-}
-
-impl<'a, 'b> NodeHandlerContext<'a, 'b> {
-    #[doc(hidden)]
-    pub fn new(state: &'a mut State, snapshot: &'b Snapshot) -> NodeHandlerContext<'a, 'b> {
-        NodeHandlerContext {
-            state: state,
-            snapshot: snapshot,
-            txs: Vec::new(),
-        }
-    }
-
-    #[doc(hidden)]
-    pub fn transactions(self) -> Vec<Box<Transaction>> {
-        self.txs
-    }
-}
-
-impl<'a, 'b> ServiceContext for NodeHandlerContext<'a, 'b> {
-    fn validator_state(&self) -> &Option<ValidatorState> {
-        self.state.validator_state()
-    }
-
-    fn snapshot(&self) -> &Snapshot {
-        self.snapshot
-    }
-
-    fn height(&self) -> Height {
-        self.state.height()
-    }
-
-    fn round(&self) -> Round {
-        self.state.round()
-    }
-
-    fn validators(&self) -> &[ValidatorKeys] {
-        self.state.validators()
-    }
-
-    fn public_key(&self) -> &PublicKey {
-        self.state.service_public_key()
-    }
-
-    fn secret_key(&self) -> &SecretKey {
-        self.state.service_secret_key()
-    }
-
-    fn actual_consensus_config(&self) -> &ConsensusConfig {
-        self.state.consensus_config()
-    }
-
-    fn actual_service_config(&self, service: &Service) -> &Value {
-        let name = service.service_name();
-        self.state.services_config().get(name).unwrap()
-    }
-
-    fn add_transaction(&mut self, tx: Box<Transaction>) {
-        assert!(tx.verify());
-        self.txs.push(tx);
-    }
-}
-
-impl<'a, 'b> fmt::Debug for NodeHandlerContext<'a, 'b> {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        write!(
-            f,
-            "ServiceContext(state: {:?}, txs: {:?})",
-            self.state,
-            self.txs
-        )
-    }
-}
-
 #[derive(Debug, Default)]
 pub struct ApiNodeState {
     incoming_connections: HashSet<SocketAddr>,
